chore: move netlifyRequest to netlify.ts
