+++ prerender-opengraph/wrangler.toml
@@ -1,5 +1,5 @@
 name = "prerender-opengraph"
-main = "src/netlify.ts"
+main = "src/index.ts"
 compatibility_date = "2022-06-24"
 routes = [
   "kodadot.xyz/rmrk/collection/*",
